Add first working, tested version of rtmptee
rtmptee manages video transcoding processes.

It watches an nginx-rtmp server for active streams.  For each active
stream, it spawns a source process, one or more sink processes, and
forwards data from each source to its sinks.  When the stream stops, it
kills the sources and sinks.  When a sink dies or stalls, it restarts
it.  When a source dies or stalls, it restarts it and its sinks.

This is a combination of 75 commits. Oo  These include two major
rewrites and dozens of introduced and resolved issues.